GAP-2396: WIP - tests
